Integracion interfaz abm usuarios con vista rrhh
