+++ .idea/deploymentTargetDropDown.xml
@@ -5,7 +5,7 @@
       <entry key="MainActivity">
         <State />
       </entry>
-      <entry key="PhotoRegisterActivity">
+      <entry key="RRHHActivity">
         <State />
       </entry>
       <entry key="RegisterActivity">
